LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/offsetProps-001-expected.txt

p
ppp pppp pppp pppp pppp pppp pppp pppp pppp pppp pppp pppp pppp

FAIL outer assert_equals: expected 292 but got 72
FAIL middle assert_equals: expected 150 but got 72
PASS inner1
FAIL inner2 assert_equals: expected 110 but got 72
